+++ src/ContaoManager/Plugin.php
@@ -1,7 +1,7 @@
 <?php
 
 /*
- * Copyright (c) 2019 Heimrich & Hannot GmbH
+ * Copyright (c) 2020 Heimrich & Hannot GmbH
  *
  * @license LGPL-3.0-or-later
  */
@@ -56,7 +56,7 @@ class Plugin implements BundlePluginInterface, ExtensionPluginInterface, ConfigP
     }
 
     /**
-     * @inheritDoc
+     * {@inheritdoc}
      */
     public function registerContainerConfiguration(LoaderInterface $loader, array $managerConfig)
     {